requirements.txt
aiofiles==24.1.0
aiogram==3.13.1
aiohappyeyeballs==2.6.1
aiohttp==3.10.5
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.3.0
cachetools==6.2.0
certifi==2025.8.3
charset-normalizer==3.4.3
distro==1.9.0
et_xmlfile==2.0.0
frozenlist==1.7.0
fuzzywuzzy==0.18.0
google-auth==2.41.1
google-genai==1.40.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
Levenshtein==0.25.1
magic-filter==1.0.12
multidict==6.6.4
numpy==2.3.3
openpyxl==3.1.5
pandas==2.2.2
propcache==0.3.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-Levenshtein==0.25.1
pytz==2025.2
RapidFuzz==3.14.1
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
tenacity==9.1.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1